Update 05/04/22
Haven't had as much time as anticipated to work on the list. Thought I'd be going back to working 5 days a week instead of 6-7. Turns out I'm currently back to 6-7 days which makes it very hard for me to also have time for this (and to do other life things). Still maintaining list, just not as quick as desired. Today's update adds multiple domains that target android devices, usually by way of malicious ad redirection